Do not whine when imagej.app.directory is unset for non-IJ1 projects
At the moment, imagej-maven-plugin is enabled by default for all
projects based on pom-scijava, even ones that are no ImageJ 1.x plugins.
That means that we warn when imagej.app.directory is not set even for
projects that will never want that.

To counter that, look at the direct dependencies and if ImageJ 1.x is
none of them (the artifactId 'ij' or 'imagej' is the tell-tale), let's
be quiet.

However, if the property *is* set and the directory does *not* exist,
let's still complain loudly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>