CopyJars: check plugin status before groupId
By checking groupId first for OME/LOCI special cases, the
bio-formats_plugins will be placed in jars/bio-formats - although should
be in /plugins, as it is an IJ1 plugin.

Thus we check plugin status first, then groupId, and if neither of those
are hits we place the target in /jars.